pyproject.toml
[project]
name = "web-aware-rag-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.0",
    "aiosqlite>=0.21.0",
    "arq>=0.26.3",
    "beautifulsoup4>=4.14.2",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.119.0",
    "groq>=0.32.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.1",
    "readability-lxml>=0.8.4.1",
    "sentence-transformers>=5.1.1",
    "tqdm>=4.67.1",
    "uvicorn[standard]>=0.37.0",
]
